Starting to get it to work

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Drive.java
@@ -4,7 +4,6 @@ import com.arcrobotics.ftclib.command.Command;
 import com.arcrobotics.ftclib.command.CommandOpMode;
 import com.arcrobotics.ftclib.command.InstantCommand;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-import com.qualcomm.robotcore.hardware.HardwareMap;
 
 @TeleOp
 public class Drive extends CommandOpMode {
@@ -14,7 +13,7 @@ public class Drive extends CommandOpMode {
         Test robot = new Test("tele", hardwareMap, telemetry);
 
         Command driveCommand = new InstantCommand(() -> {
-            robot.driveSubsystem.teleDrive(() -> gamepad1.left_stick_x, () -> gamepad1.left_stick_y, ()-> gamepad1.right_stick_y, .5);
+            robot.driveSubsystem.teleDrive(() -> (double) gamepad1.left_stick_x, () -> (double) gamepad1.left_stick_y, ()-> (double) gamepad1.right_stick_y, .5);
         });
 
         schedule(driveCommand);
